fix navio_sysfs_pwm_out: use correct type for buflen (required by MultirotorMixer::from_text)

+++ src/drivers/navio_sysfs_pwm_out/navio_sysfs_pwm_out.cpp
@@ -124,7 +124,7 @@ int mixer_control_callback(uintptr_t handle,
 int initialize_mixer(const char *mixer_filename)
 {
 	char buf[2048];
-	size_t buflen = sizeof(buf);
+	unsigned buflen = sizeof(buf);
 
 	PX4_INFO("Trying to initialize mixer from config file %s", mixer_filename);
